+++ schemes/bgv/evaluator.go
@@ -1544,7 +1544,7 @@ func (eval Evaluator) InnerSum(ctIn *rlwe.Ciphertext, batchSize, n int, opOut *r
 			return
 		}
 
-		if err = eval.Evaluator.PartialTrace(ctIn, batchSize, n/2, opOut); err != nil {
+		if err = eval.Evaluator.PartialTracesSum(ctIn, batchSize, n/2, opOut); err != nil {
 			return
 		}
 
@@ -1561,7 +1561,7 @@ func (eval Evaluator) InnerSum(ctIn *rlwe.Ciphertext, batchSize, n int, opOut *r
 		return
 	}
 
-	err = eval.Evaluator.PartialTrace(ctIn, batchSize, n, opOut)
+	err = eval.Evaluator.PartialTracesSum(ctIn, batchSize, n, opOut)
 	return
 }
 
@@ -1581,7 +1581,7 @@ func (eval Evaluator) InnerSum(ctIn *rlwe.Ciphertext, batchSize, n int, opOut *r
 //
 // Calling RotateAndAdd(ctIn, 1, n, opOut) can be used to compute the inner sum of the first n slots of a plaintext.
 func (eval Evaluator) RotateAndAdd(ctIn *rlwe.Ciphertext, batchSize, n int, opOut *rlwe.Ciphertext) (err error) {
-	err = eval.Evaluator.PartialTrace(ctIn, batchSize, n, opOut)
+	err = eval.Evaluator.PartialTracesSum(ctIn, batchSize, n, opOut)
 	return
 }